[BOSK-88] Update block executors, add group blacklisting
